requirements.txt
absl-py==0.9.0
aiohttp==3.7.4.post0
alembic==1.8.0
appdirs==1.4.4
appnope==0.1.0
asgiref==3.2.7
astroid==2.4.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
Automat==20.2.0
awscli==1.25.23
backcall==0.1.0
bcrypt==3.2.2
beautifulsoup4==4.9.1
bleach==3.1.5
blinker==1.4
boto==2.49.0
boto3==1.24.23
botocore==1.27.23
bs4==0.0.1
cachelib==0.9.0
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.15.0
chardet==3.0.4
click==8.1.3
colorama==0.4.4
constantly==15.1.0
cryptography==37.0.2
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
discord==1.7.3
discord.py==1.7.3
distlib==0.3.0
Django==3.0.6
django-allauth==0.42.0
django-countries==7.0
django-crispy-forms==1.9.1
django-filter==2.3.0
djangorestframework==3.11.1
dnspython==2.2.1
docutils==0.16
email-validator==1.2.1
entrypoints==0.3
feedwerk==1.1.0
filelock==3.0.12
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Blogging==2.0.0.post2
Flask-Caching==2.0.0
Flask-CKEditor==0.4.6
Flask-DebugToolbar==0.11.0
Flask-Email==1.4.4
Flask-FileUpload==0.5.0
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
future==0.18.2
gast==0.3.3
google-auth==1.16.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.29.0
gunicorn==20.1.0
h11==0.8.1
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
http3==0.6.7
huepy==1.2.1
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==4.11.4
importlib-resources==5.8.0
incremental==21.3.0
iniconfig==1.0.0
instabot==0.117.0
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
isort==4.3.21
itemadapter==0.6.0
itemloaders==1.0.4
itsdangerous==2.1.2
jedi==0.17.0
Jinja2==3.1.2
jmespath==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==2.1.4
jupyterlab-server==1.1.5
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.1
Mako==1.2.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
mock==4.0.2
more-itertools==8.4.0
moviepy==1.0.3
multidict==5.2.0
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.4
oauthlib==3.1.0
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.4
pandocfilters==1.4.2
Paperboy==1.0.1
parsel==1.6.0
parso==0.7.0
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
pipenv==2020.6.2
pluggy==0.13.1
proglog==0.1.9
prometheus-client==0.8.0
prompt-toolkit==3.0.5
Protego==0.2.1
protobuf==3.12.2
psycopg2-binary==2.9.3
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
Pygments==2.6.1
PyJWT==2.3.0
pylint==2.5.2
pylint-django==2.3.0
pylint-plugin-utils==0.6
pynput==1.7.5
pyobjc-core==8.0
pyobjc-framework-ApplicationServices==8.0
pyobjc-framework-Cocoa==8.0
pyobjc-framework-Quartz==8.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
pytest==6.0.0
python-dateutil==2.8.2
python-dotenv==0.19.1
python-editor==1.0.4
python-slugify==6.1.2
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.4.1
pyzmq==19.0.1
queuelib==1.6.2
requests==2.23.0
requests-async @ https://github.com/encode/requests-async/archive/master.zip
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
responses==0.10.15
rfc3986==1.4.0
rsa==4.7.2
s3transfer==0.6.0
schedule==0.6.0
scipy==1.4.1
Scrapy==2.6.1
selenium==3.141.0
Send2Trash==1.5.0
service-identity==21.1.0
shortuuid==1.0.9
six==1.16.0
soupsieve==2.0.1
SQLAlchemy==1.4.37
sqlparse==0.3.1
stripe==2.55.2
svgwrite==1.4
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
tldextract==3.3.0
toml==0.10.1
tornado==6.0.4
tqdm==4.48.0
traitlets==4.3.3
Tree==0.2.4
twilio==7.9.0
Twisted==22.4.0
typing-extensions==3.10.0.2
urllib3==1.26.9
virtualenv==20.0.21
virtualenv-clone==0.5.4
w3lib==1.22.0
wcwidth==0.2.3
webencodings==0.5.1
Werkzeug==2.1.2
wrapt==1.12.1
WTForms==3.0.1
yarl==1.7.0
zipp==3.8.0
zope.interface==5.4.0
